refactor: Split ProjectModel into separate classes for better separation
Refactored the monolithic ProjectModel into three separate classes:
- AbstractProjectModel: Base class with common functionality
- StarterProjectModel: For start.vaadin.com/skeleton projects
- HelloWorldProjectModel: For start.vaadin.com/helloworld projects

This improves code organization and makes it clearer which options
apply to which project type. Each model now only contains the fields
and methods relevant to its specific project type.

Changes:
- Created AbstractProjectModel base class with common methods
- Created StarterProjectModel with prerelease and framework options
- Created HelloWorldProjectModel with language, build tool, and architecture options
- Updated VaadinProjectWizardPage to use appropriate model based on selection
- Updated NewVaadinProjectWizard to use AbstractProjectModel interface
- Created separate test classes for each model type
- Removed old monolithic ProjectModel class
